requirements.txt
streamlit
PyPDF2
langchain
python-dotenv
google-generativeai
Pillow
pytesseract

